New Screen instead of popup for Game Parameters, during editing/creating scenario.
